% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-examples.R
\name{aigra_example_item_bank}
\alias{aigra_example_item_bank}
\title{Get path to bundled AIGRA example item bank}
\usage{
aigra_example_item_bank()
}
\value{
Path to the bundled example CSV item bank.
}
\description{
Get path to bundled AIGRA example item bank
}
