% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-api.R
\name{aigra_latest_output}
\alias{aigra_latest_output}
\title{Get latest AIGRA CSV output path}
\usage{
aigra_latest_output()
}
\value{
Path to the latest AIGRA CSV output.
}
\description{
Get latest AIGRA CSV output path
}
