% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-multimodal-template.R
\name{aigra_multimodal_template_items}
\alias{aigra_multimodal_template_items}
\title{Create AIGRA multimodal template items}
\usage{
aigra_multimodal_template_items(diagram_dir = NULL, create_diagrams = TRUE)
}
\arguments{
\item{diagram_dir}{Directory where source diagrams are stored or should be created.}

\item{create_diagrams}{If TRUE, creates sample source diagram PNG files.}
}
\value{
A data frame containing multimodal sample items.
}
\description{
Create AIGRA multimodal template items
}
