% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-outputs.R
\name{aigra_outputs}
\alias{aigra_outputs}
\title{List AIGRA output files}
\usage{
aigra_outputs(pattern = "aigra(_tabular)?_results_.*\\\\.(csv|jsonl)$")
}
\arguments{
\item{pattern}{File-name pattern.}
}
\value{
A data frame of output files.
}
\description{
List AIGRA output files
}
