% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-diagrams.R
\name{aigra_plot_diagram}
\alias{aigra_plot_diagram}
\title{Plot an AIGRA diagram image}
\usage{
aigra_plot_diagram(image_path)
}
\arguments{
\item{image_path}{Path to a PNG image.}
}
\value{
Invisibly returns the image path.
}
\description{
Plot an AIGRA diagram image
}
