% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-validation.R
\name{aigra_print_validation}
\alias{aigra_print_validation}
\title{Print tabular item-bank validation results}
\usage{
aigra_print_validation(validation)
}
\arguments{
\item{validation}{Validation result returned by \code{\link[=aigra_validate_tabular_items]{aigra_validate_tabular_items()}}.}
}
\value{
Invisibly returns the validation object.
}
\description{
Print tabular item-bank validation results
}
