% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-python.R
\name{aigra_python_info}
\alias{aigra_python_info}
\title{Show AIGRA Python environment information}
\usage{
aigra_python_info(backend_path = NULL)
}
\arguments{
\item{backend_path}{Path to the AIGRA_BACKEND folder.}
}
\value{
A list with Python environment information.
}
\description{
Show AIGRA Python environment information
}
