% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-outputs.R
\name{aigra_read_latest_csv}
\alias{aigra_read_latest_csv}
\title{Read latest AIGRA CSV output}
\usage{
aigra_read_latest_csv()
}
\value{
A data frame containing the latest AIGRA CSV output.
}
\description{
Read latest AIGRA CSV output
}
