% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigra-api.R
\name{aigra_read_latest_output}
\alias{aigra_read_latest_output}
\title{Read latest AIGRA CSV output}
\usage{
aigra_read_latest_output()
}
\value{
A data frame containing the latest AIGRA CSV output.
}
\description{
Read latest AIGRA CSV output
}
